refactor(webapp): move auth store to runes
